Bm branch

+++ Web/routes/api/user.php
@@ -8,18 +8,20 @@ use Illuminate\Http\Request;
 |--------------------------------------------------------------------------
 */
 
+Route::post('/check_login', 'UserController@login')->middleware('api');
+
+Route::post('/logout', 'UserController@logout')->middleware('api');
+
 Route::group(['prefix' => 'user', 'middleware' => ['api']], function () {
     Route::get('/', 'UserController@index');
 
     Route::post('/', 'UserController@store');
 
-    Route::post('/login', 'UserController@login');
+    Route::get('/{id}', 'UserController@show')->where('id', '[0-9]+')->middleware('auth');
 
-    Route::get('/{id}', 'UserController@show')->where('id', '[0-9]+');
+    Route::patch('/{id}', 'UserController@update')->where('id', '[0-9]+')->middleware('auth');
 
-    Route::patch('/{id}', 'UserController@update')->where('id', '[0-9]+');
-
-    Route::delete('/{id}', 'UserController@destroy')->where('id', '[0-9]+');
+    Route::delete('/{id}', 'UserController@destroy')->where('id', '[0-9]+')->middleware('auth');
 
     Route::get('/search/{text}', 'UserController@search');
